Implement and use template function for JSON.
We build GraphQL queries with Go templates. We were using a built-in
`js` template function for escaping arbitrary strings before using them
as JSON strings. Which _almost_ works; it's not the intended use case
for `js`. We actually needed a custom template function, as implemented
here.

Most of the changes in this commit are simply replacing `js` with
`json` where templates are defined. The meat of the commit is in
`query.go`.